fix: 4 more bugs found during beta testing
- api.py: freeze text encoder by default for LoRA (train_text_encoder=False)
  fixes TE running every step (no cache) and being saved in checkpoints wastefully
- train_backend_zimage.py: try manual loading FIRST for single-file checkpoints
  avoids downloading 12+ GB transformer weights from HF just to replace with local
  checkpoint (local file has transformer, only TE+VAE need HF download)
- train_backend_base.py: cast fp16 LoRA params to float32 for GradScaler compat
- trainer.py: disable FP8 training when LoRA is active (PEFT FP8LinearWrapper lacks .weight)
- optimizer_factory.py: normalize param_groups lr for Prodigy/DAdaptAdam
- optimizer_factory.py: _CombinedOptimizer inherits torch.optim.Optimizer (LRScheduler compat)

+++ dataset_sorter/optimizers.py
@@ -206,7 +206,7 @@ class Marmotte(Optimizer):
                 # ── Gradient-norm adaptive scaling ────────────────────
                 grad_rms = grad.norm() / max(math.sqrt(grad.numel()), 1.0)
                 ema = state["grad_rms_ema"]
-                ema.lerp_(grad_rms, 1 - rms_beta)
+                ema.lerp_(grad_rms.float(), 1 - rms_beta)
                 ema_corrected = ema / (1 - rms_beta ** step_num)
                 if ema_corrected > eps:
                     adaptive_scale = (grad_rms / ema_corrected).clamp(0.1, 10.0).item()
